pyproject.toml
[project]
name = "spotify-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.5",
    "jinja2>=3.1.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "spotipy>=2.24.0",
]
